Renaming voting.hpp to generator.hpp since it only holds the generator class.
