+++ nano/node/voting/generator.cpp
@@ -6,7 +6,7 @@
 #include <nano/node/transport/inproc.hpp>
 #include <nano/node/voting/history.hpp>
 #include <nano/node/voting/processor.hpp>
-#include <nano/node/voting/voting.hpp>
+#include <nano/node/voting/generator.hpp>
 #include <nano/node/wallet.hpp>
 #include <nano/secure/ledger.hpp>
 #include <nano/store/component.hpp>
