+++ nano/node/voting/generator.hpp
